Removed lightslider.js and made a very lightweight slider
Jquery lightslider removed (it had very much bugs) and changed the
behaviour of the thumbimages. Thumbimages are now on one row in
list-items. When an image becomes 'active' meaning it is on the large
view, the UL element is slided to an offset of the position of the
'active' image and the images before it. This way the active image is
usually the second image visible in the bottom bar. Exception is when it
is the first image.